+++ autoware_utils_geometry/CMakeLists.txt
@@ -1,0 +1,10 @@
+cmake_minimum_required(VERSION 3.14)
+project(autoware_utils_geometry)
+
+find_package(autoware_cmake REQUIRED)
+autoware_package()
+
+if(BUILD_TESTING)
+endif()
+
+ament_auto_package()
